+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 #### Additions
 * LF/AD-2>2.19 (AD navigational aids relevant to VFR)
 
+#### Changes
+* Renamed `~/.aipp/<region>/archive` to `~/.aipp/<region>/sources`
+
 ## 0.2.4
 
 #### Additions